requirements.txt
gradio
ffmpeg-python
pydub
Pillow
PyPDF2
python-docx
